+++ example/spec/parallel/ingestGranule/IngestUMMGSuccessSpec.js
@@ -35,7 +35,7 @@ const {
   getTEARequestHeaders,
 } = require('@cumulus/integration-tests/api/distribution');
 const { LambdaStep } = require('@cumulus/integration-tests/sfnStep');
-const { encodedConstructCollectionId } = require('../../helpers/Collections');
+const { constructCollectionId } = require('@cumulus/message/Collections');
 
 const {
   loadConfig,
@@ -131,7 +131,7 @@ describe('The S3 Ingest Granules workflow configured to ingest UMM-G', () => {
 
       collection = { name: `MOD09GQ${testSuffix}`, version: '006' };
       provider = { id: `s3_provider${testSuffix}` };
-      const newCollectionId = encodedConstructCollectionId(collection.name, collection.version);
+      const newCollectionId = constructCollectionId(collection.name, collection.version);
 
       process.env.system_bucket = config.bucket;
 
@@ -215,7 +215,7 @@ describe('The S3 Ingest Granules workflow configured to ingest UMM-G', () => {
     await removePublishedGranule({
       prefix: config.stackName,
       granuleId: inputPayload.granules[0].granuleId,
-      collectionId: encodedConstructCollectionId(collection.name, collection.version),
+      collectionId: constructCollectionId(collection.name, collection.version),
     });
     await apiTestUtils.deletePdr({
       prefix: config.stackName,
@@ -517,7 +517,7 @@ describe('The S3 Ingest Granules workflow configured to ingest UMM-G', () => {
       const moveGranuleResponse = await moveGranule({
         prefix: config.stackName,
         granuleId: inputPayload.granules[0].granuleId,
-        collectionId: encodedConstructCollectionId(collection.name, collection.version),
+        collectionId: constructCollectionId(collection.name, collection.version),
         destinations,
       });