+++ example/spec/parallel/syncGranule/SyncGranuleDuplicateHandlingSpec.js
@@ -16,7 +16,7 @@ const {
   cleanupProviders,
 } = require('@cumulus/integration-tests');
 const { getExecutionUrlFromArn } = require('@cumulus/message/Executions');
-const { encodedConstructCollectionId } = require('../../helpers/Collections');
+const { constructCollectionId } = require('@cumulus/message/Collections');
 
 const { buildAndExecuteWorkflow } = require('../../helpers/workflowUtils');
 const {
@@ -79,7 +79,7 @@ describe('When the Sync Granule workflow is configured', () => {
 
     collection = { name: `MOD09GQ${testSuffix}`, version: '006' };
     provider = { id: `s3_provider${testSuffix}` };
-    const newCollectionId = encodedConstructCollectionId(collection.name, collection.version);
+    const newCollectionId = constructCollectionId(collection.name, collection.version);
 
     // populate collections, providers and test data
     await Promise.all([
@@ -139,7 +139,7 @@ describe('When the Sync Granule workflow is configured', () => {
         await waitForGranuleAndDelete(
           config.stackName,
           granule.granuleId,
-          encodedConstructCollectionId(collection.name, collection.version),
+          constructCollectionId(collection.name, collection.version),
           ['completed', 'failed']
         );
       }