+++ example/spec/parallel/testAPI/bulkOperationsSuccessSpec.js
@@ -26,7 +26,7 @@ const {
 const { getGranuleWithStatus } = require('@cumulus/integration-tests/Granules');
 const { createProvider } = require('@cumulus/integration-tests/Providers');
 const { createOneTimeRule } = require('@cumulus/integration-tests/Rules');
-const { encodedConstructCollectionId } = require('../../helpers/Collections');
+const { constructCollectionId } = require('@cumulus/message/Collections');
 
 const {
   createTimestampedTestId,
@@ -102,7 +102,7 @@ describe('POST /granules/bulk', () => {
           Body: 'asdf',
         });
 
-        collectionId = encodedConstructCollectionId(collection.name, collection.version);
+        collectionId = constructCollectionId(collection.name, collection.version);
         granuleId = randomId('granule-id-');
         console.log('granuleId', granuleId);