+++ example/spec/serial/ingestPdrWithNodeNameFtpSpec.js
@@ -23,8 +23,8 @@ const {
   waitForCompletedExecution,
 } = require('@cumulus/integration-tests');
 
+const { constructCollectionId } = require('@cumulus/message/Collections');
 const { buildFtpProvider } = require('../helpers/Providers');
-const { encodedConstructCollectionId } = require('../helpers/Collections');
 const { buildAndExecuteWorkflow } = require('../helpers/workflowUtils');
 const {
   createTestDataPath,
@@ -351,7 +351,7 @@ describe('Ingesting from PDR', () => {
           // delete ingested granule(s)
           await Promise.all(
             finalOutput.payload.granules.map(async (g) => {
-              const newCollectionId = encodedConstructCollectionId(addedCollection.name, addedCollection.version);
+              const newCollectionId = constructCollectionId(addedCollection.name, addedCollection.version);
               await waitForApiStatus(
                 getGranule,
                 {